+++ flutter_application_1/lib/timer.dart
@@ -1,0 +1,105 @@
+import 'package:flutter/material.dart';
+import 'dart:async';
+
+class TimerPage extends StatefulWidget {
+  const TimerPage({super.key});
+
+  @override
+  _TimerPageState createState() => _TimerPageState();
+}
+
+class _TimerPageState extends State<TimerPage> {
+  Timer? _timer;
+  int _seconds = 0;
+  bool _isRunning = false;
+
+  void _toggleTimer() {
+    if (_isRunning) {
+      _stopTimer();
+    } else {
+      _startTimer();
+    }
+  }
+
+  void _startTimer() {
+    setState(() {
+      _isRunning = true;
+    });
+    _timer = Timer.periodic(const Duration(seconds: 1), (timer) {
+      setState(() {
+        _seconds++;
+      });
+    });
+  }
+
+  void _stopTimer() {
+    setState(() {
+      _isRunning = false;
+    });
+    _timer?.cancel();
+  }
+
+  void _resetTimer() {
+    _stopTimer(); // Stop the timer first
+    setState(() {
+      _seconds = 0; // Reset the seconds to 0
+    });
+  }
+
+  String _formatTime(int totalSeconds) {
+    int hours = totalSeconds ~/ 3600;
+    int minutes = (totalSeconds % 3600) ~/ 60;
+    int seconds = totalSeconds % 60;
+
+    // Format the output to display as HH:MM:SS
+    String hoursStr = hours.toString().padLeft(2, '0');
+    String minutesStr = minutes.toString().padLeft(2, '0');
+    String secondsStr = seconds.toString().padLeft(2, '0');
+
+    return '$hoursStr:$minutesStr:$secondsStr';
+  }
+
+  
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Timer Page'),
+      ),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            const Text(
+              'Timer is running for:',
+            ),
+            Text(
+              _formatTime(_seconds),
+              style: Theme.of(context).textTheme.headlineMedium,
+            ),
+          ],
+        ),
+      ),
+      floatingActionButton: Column(
+        mainAxisAlignment: MainAxisAlignment.end,
+        children: [
+
+          FloatingActionButton(
+            onPressed: _toggleTimer,
+            tooltip: _isRunning ? 'Stop Timer' : 'Start Timer',
+            child: Icon(_isRunning ? Icons.pause : Icons.play_arrow),
+          ),
+
+          const SizedBox(height: 16), // Add some space between the buttons
+          FloatingActionButton(
+            onPressed: _resetTimer,
+            tooltip: 'Restart Timer',
+            child: const Icon(Icons.replay),
+          ),
+
+        ],
+      ),
+    );
+  }
+}
